// editor/src/flow/types/FunctionType.ts
export enum FunctionType {
  And = 'And',

  Average = 'Average',

  Calculator = 'Calculator',

  Calendar = 'Calendar',

  Clamp = 'Clamp',

  Comparator = 'Comparator',

  Delay = 'Delay',

  If = 'If',

  Invert = 'Invert',

  // Latches a value for later reference
  Latch = 'Latch',

  Max = 'Max',

  Min = 'Min',

  Or = 'Or',

  Override = 'Override',

  Pid = 'Pid',

  Pulse = 'Pulse',

  Schedule = 'Schedule',

  Selector = 'Selector',

  // Sequences a set of outputs based on input selectors
  Sequence = 'Sequence',

  // Spans an input value range to an output value range
  Span = 'Span',

  Split = 'Split',

  Timer = 'Timer',

  Xnor = 'Xnor',

  Xor = 'Xor'
}